Refactored `client` code into view and model classes (out of auth) Extended event mechanism to support user document uploads (TN-692, TN-797)
