+++ portal/config/gil/Intervention.json
@@ -5,6 +5,7 @@
       "description": "Assessment Engine",
       "name": "assessment_engine",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -14,6 +15,7 @@
       "link_label": "Care Plan",
       "name": "care_plan",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -22,6 +24,7 @@
       "display_rank": 10,
       "name": "community_of_wellness",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -31,6 +34,7 @@
       "link_label": "Go to Decision Support",
       "name": "decision_support_p3p",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -40,6 +44,7 @@
       "link_label": "Not Available",
       "name": "decision_support_unavailable",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -49,6 +54,7 @@
       "link_label": "WiserCare",
       "name": "decision_support_wisercare",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -57,6 +63,7 @@
       "name": "default",
       "public_access": false,
       "resourceType": "Intervention",
+      "subscribed_events": 1,
       "status_text": ""
     },
     {
@@ -66,6 +73,7 @@
       "link_label": "Exercise and Diet",
       "name": "exercise_diet",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -75,6 +83,7 @@
       "link_label": "lived experience",
       "name": "lived_experience",
       "public_access": true,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -84,6 +93,7 @@
       "link_label": "Go to Symptom Tracker",
       "name": "self_management",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
@@ -94,6 +104,7 @@
       "name": "sexual_recovery",
       "public_access": false,
       "resourceType": "Intervention",
+      "subscribed_events": 1,
       "status_text": ""
     },
     {
@@ -101,12 +112,14 @@
       "description": "Social Support Network",
       "name": "social_support",
       "public_access": false,
+      "subscribed_events": 1,
       "resourceType": "Intervention"
     },
     {
       "description": "MUSIC Integration",
       "name": "music",
       "public_access": false,
+      "subscribed_events": 3,
       "resourceType": "Intervention"
     }
   ],
