+++ portal/gil/templates/gil/base.html
@@ -78,7 +78,7 @@
           {% endif %}
           <li class="side-nav-items__item side-nav-items__item--about"><a href="{{url_for('gil.about')}}">{{_("About TrueNTH")}}</a></li>
           {% if user and user.has_role(ROLE.APPLICATION_DEVELOPER) %}
-              <li class="side-nav-items__item side-nav-items__item--clients"><a href="{{url_for('auth.clients_list')}}">{{_("Client Applications")}}</a></li>
+              <li class="side-nav-items__item side-nav-items__item--clients"><a href="{{url_for('client.clients_list')}}">{{_("Client Applications")}}</a></li>
           {% endif %}
           {% if user and (user.has_role(ROLE.ADMIN) or user.has_role(ROLE.ANALYST)) %}
               <li class="side-nav-items__item side-nav-items__item--clients"><a href="{{url_for('portal.reporting_dashboard')}}">{{_("Reporting Dashboard")}}</a></li>
