+++ portal/views/crossdomain.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from flask import make_response, request, current_app
 from functools import update_wrapper
 
-from ..models.auth import validate_origin
+from ..models.client import validate_origin
 
 
 def crossdomain(
